+++ library/src/main/kotlin/com/devbrackets/android/playlistcore/api/MediaPlayerApi.kt
@@ -2,9 +2,10 @@ package com.devbrackets.android.playlistcore.api
 
 import android.support.annotation.FloatRange
 import android.support.annotation.IntRange
-import com.devbrackets.android.playlistcore.event.MediaProgress
+import com.devbrackets.android.playlistcore.data.MediaProgress
 import com.devbrackets.android.playlistcore.listener.MediaStatusListener
 
+//TODO: maybe just call this a MediaPlayer and ignore the overlap with the Android mediaPlayer class?
 interface MediaPlayerApi<I : PlaylistItem> {
     /**
      * Determines if media is currently playing on the
