+++ library/src/main/kotlin/com/devbrackets/android/playlistcore/components/mediacontrols/MediaControlsProvider.kt
@@ -1,7 +1,7 @@
-package com.devbrackets.android.playlistcore.helper.mediacontrols
+package com.devbrackets.android.playlistcore.components.mediacontrols
 
 import android.support.v4.media.session.MediaSessionCompat
-import com.devbrackets.android.playlistcore.helper.notification.MediaInfo
+import com.devbrackets.android.playlistcore.data.MediaInfo
 
 interface MediaControlsProvider {
     fun update(mediaInfo: MediaInfo, mediaSession: MediaSessionCompat)
