+++ library/src/main/kotlin/com/devbrackets/android/playlistcore/listener/PlaylistListener.kt
@@ -18,7 +18,7 @@ package com.devbrackets.android.playlistcore.listener
 
 import com.devbrackets.android.playlistcore.api.PlaylistItem
 import com.devbrackets.android.playlistcore.service.BasePlaylistService
-import com.devbrackets.android.playlistcore.service.PlaybackState
+import com.devbrackets.android.playlistcore.data.PlaybackState
 
 /**
  * A simple callback interface for listening to [BasePlaylistService]
