+++ library/src/main/kotlin/com/devbrackets/android/playlistcore/listener/ProgressListener.kt
@@ -16,7 +16,7 @@
 
 package com.devbrackets.android.playlistcore.listener
 
-import com.devbrackets.android.playlistcore.event.MediaProgress
+import com.devbrackets.android.playlistcore.data.MediaProgress
 
 interface ProgressListener {
