+++ library/src/main/kotlin/com/devbrackets/android/playlistcore/manager/BasePlaylistManager.kt
@@ -24,14 +24,14 @@ import android.support.annotation.IntRange
 import android.util.Log
 import com.devbrackets.android.playlistcore.api.MediaPlayerApi
 import com.devbrackets.android.playlistcore.api.PlaylistItem
-import com.devbrackets.android.playlistcore.event.MediaProgress
-import com.devbrackets.android.playlistcore.event.PlaylistItemChange
+import com.devbrackets.android.playlistcore.data.MediaProgress
+import com.devbrackets.android.playlistcore.data.PlaylistItemChange
 import com.devbrackets.android.playlistcore.listener.PlaylistListener
 import com.devbrackets.android.playlistcore.listener.ProgressListener
 import com.devbrackets.android.playlistcore.listener.ServiceListener
 import com.devbrackets.android.playlistcore.service.BasePlaylistService
-import com.devbrackets.android.playlistcore.service.PlaybackState
-import com.devbrackets.android.playlistcore.helper.playlist.PlaylistHandler
+import com.devbrackets.android.playlistcore.data.PlaybackState
+import com.devbrackets.android.playlistcore.components.playlisthandler.PlaylistHandler
 import com.devbrackets.android.playlistcore.service.RemoteActions
 import java.lang.ref.WeakReference
 import java.util.*
